feat: add support for breaking change token in the message footer
Extends support for signaling breaking changes in commit messages according to the
[Conventional Commits Specification](https://www.conventionalcommits.org/en/v1.0.0/#specification)
(see points 11, 12, 13, 15 and 16).

Previously, only the suffix "!" after the type/scope was recognized as a
breaking change. With this change, the footer "BREAKING CHANGE" or "BREAKING-CHANGE" will also be
identified (as per point 12 and 16 of the specification), these are NOT case-sensitive
(as per point 15 of the specification).

The footer `BREAKING CHANGE` can be omitted as per point 13 of the specification.

Additionally, the description given in `BREAKING CHANGE` if present instead of the commit
description as per point 12 of the specification

For reference these are the points of the specifications that are being implemented:

11. Breaking changes MUST be indicated in the type/scope prefix of a commit, or as an entry in the
    footer.
12. If included as a footer, a breaking change MUST consist of the uppercase text BREAKING CHANGE,
    followed by a colon, space, and description, e.g., BREAKING CHANGE: environment variables now
    take precedence over config files.
13. If included in the type/scope prefix, breaking changes MUST be indicated by a ! immediately
    before the :. If ! is used, BREAKING CHANGE: MAY be omitted from the footer section, and the
    commit description SHALL be used to describe the breaking change.
15. The units of information that make up Conventional Commits MUST NOT be treated as case
    sensitive by implementors, with the exception of BREAKING CHANGE which MUST be uppercase.
16. BREAKING-CHANGE MUST be synonymous with BREAKING CHANGE, when used as a token in a footer.